Move things to better fitting places
